==> 01_PopulationSegmentation/01_scripts/01_Data_Extraction.ipynb <==
<b>Project</b>: Population segmentation and transition probability estimation using data on health and health-related social service needs from the US Health and Retirement Study <br>
<b>Project section</b>: Population segmentation <br>
<b>Python version</b>: 3.9.7 <br>
<b>File name</b>: 01_Data_Extraction.ipynb <br>
<b>Data required</b>: Health and Retirement Study (HRS) datasets: RAND HRS Longitudinal File 2018, Cross-Wave Tracker File, 2018 RAND HRS Fat File, 2016 RAND HRS Fat File, 2014 RAND HRS Fat File, 2012 RAND HRS Fat File, 2010 RAND HRS Fat File, 2008 RAND HRS Fat File, 2006 RAND HRS Fat File <br>
<b>Outcome</b>: Extracts necessary data from various Health and Retirement Study (HRS) datasets and prepares a dataset for segmentation <br>
<b>Author</b>: Lize Duminy<br>
<b>Date</b>: 2023.03.19 

# Instructions for use

This script requires raw data to be downloaded from the Health and Retirement Study (HRS) platform for researchers. The daw data required for segmentation are public datasets, accessable to all persons with a registered HRS user account. 

1. If you do not already have a user account on the HRS platform for researchers, follow the instruction in [User account creation userform](https://hrsdata.isr.umich.edu/user/register) to create a new user account. 
2. Log into the [HRS platform](https://hrsdata.isr.umich.edu/user/login).
3. Navigate to each of the __data product pages__ listed in the table below and download the specified __zip folders__. If any of the zip folders listed are no longer available (since the folders are incrementally replaced with newer versions), download the STATA version of the available file and update the name of the folder in the preparation section of the code in __1.2 USER INPUT REQUIRED: Change filename if any of the zip folders were updated with newer versions__. For further documentation related to each data product, see the linked __documentation__.   
5. Extract all zip folders into a folder designated for raw data on your system.
6. Replace the filepath of the variable __global_path__ in the code below (currently _C:/Users/LizeDuminy/data/HRS/_) with the filepath of your designated folder for raw data in __1.1. USER INPUT REQUIRED: Replace this filepath with the filepath of your designated folder for raw data__. 



| Data product page | Zip folder name | Documentation |
| -: | :-: | :- |
| [RAND HRS Longitudinal File 2018](https://hrsdata.isr.umich.edu/data-products/rand-hrs-longitudinal-file-2018) | __randhrs1992_2018v2_STATA.zip__ | [randhrs1992_2018v2.pdf](https://hrsdata.isr.umich.edu/sites/default/files/documentation/other/1658948491/randhrs1992_2018v2.pdf) |
| [Cross-Wave Tracker File](https://hrsdata.isr.umich.edu/data-products/cross-wave-tracker-file) | __trk2020v2.zip__ | [README_trk20.pdf](https://hrsdata.isr.umich.edu/sites/default/files/documentation/data-descriptions/1671578103/trk20.pdf) |
| [2018 RAND HRS Fat File](https://hrsdata.isr.umich.edu/data-products/2018-rand-hrs-fat-file) | __h18f2a_STATA.zip__ | [README_fat18.pdf](https://hrsdata.isr.umich.edu/sites/default/files/documentation/other/1658866817/README_fat18.pdf) |
| [2016 RAND HRS Fat File](https://hrsdata.isr.umich.edu/data-products/2016-rand-hrs-fat-file) | __h16f2b_STATA.zip__ | [README_fat16.pdf](https://hrsdata.isr.umich.edu/sites/default/files/documentation/other/1615838140/README_fat16.pdf) |
| [2014 RAND HRS Fat File](https://hrsdata.isr.umich.edu/data-products/2014-rand-hrs-fat-file) | __h14f2b_STATA.zip__ | [README_fat14.pdf](https://hrsdata.isr.umich.edu/sites/default/files/documentation/other/1615583715/README_fat14.pdf) |
| [2012 RAND HRS Fat File](https://hrsdata.isr.umich.edu/data-products/2012-rand-hrs-fat-file) | __h12f3a_STATA.zip__ | [README_fat12.pdf](https://hrsdata.isr.umich.edu/sites/default/files/documentation/other/1615837583/README_fat12.pdf) |
| [2010 RAND HRS Fat File](https://hrsdata.isr.umich.edu/data-products/2010-rand-hrs-fat-file) | __hd10f6a_STATA.zip__ | [README_fat10.pdf](https://hrsdata.isr.umich.edu/sites/default/files/documentation/other/1658943550/README_fat10.pdf) |
| [2008 RAND HRS Fat File](https://hrsdata.isr.umich.edu/data-products/2008-rand-hrs-fat-file) | __h08f3a_STATA.zip__ | [README_fat08.pdf](https://hrsdata.isr.umich.edu/sites/default/files/documentation/other/README_fat08.pdf) |
| [2006 RAND HRS Fat File](https://hrsdata.isr.umich.edu/data-products/2006-rand-hrs-fat-file) | __h06f4a_STATA.zip__ | [README_fat06.pdf](https://hrsdata.isr.umich.edu/sites/default/files/documentation/other/1658871121/README_fat06.pdf) |

# 1. Preparation 

#### Import packages

import pandas as pd
import numpy as np
import os
import os.path
import pyreadstat 

### 1.1. USER INPUT REQUIRED: Replace this filepath with the filepath of your designated folder for raw data 

#File path for daw data
global_path=os.path.join("C:/Users/LizeDuminy/data/HRS/raw_data/")

#File path for script output
global_path2=os.path.join("C:/Users/LizeDuminy/data/HRS/data/")

### 1.2. USER INPUT REQUIRED: Change filename if any of the zip folders were updated with newer versions

#RAND HRS Longitudinal File 2018
path_randhrs = os.path.join(global_path , 
                            "randhrs1992_2018v2_STATA/",
                            "randhrs1992_2018v2.dta")

#Cross-Wave Tracker File
path_trk = os.path.join(global_path , "trk2020v2/trk2020tr_r.dta")

#RAND HRS Fat File
path_hrsCore = {
    14: "h18f2a_STATA/h18f2a.dta",
    13: "h16f2b_STATA/h16f2b.dta",
    12: "h14f2b_STATA/h14f2b.dta", 
    11: "h12f3a_STATA/h12f3a.dta", 
    10: "hd10f6a_STATA/hd10f6a.dta", 
    9:  "h08f3a_STATA/h08f3a.dta", 
    8:  "h06f4a_STATA/h06f4a.dta",
    }

# 2. RAND HRS Longitudinal File 2018

This section extracts and transforms the __RAND HRS Longitudinal File 2018__.

#### Load RAND HRS data as "df_randhrs"

df_randhrs = pd.read_stata(path_randhrs)

#### Define columns that have a wave-specific pre-fix on individual level

variables = {
    #chronic conditions
    'hibp',     #High blood pressure
    'diab',     #Diabetes
    'cancr',    #Cancer (excl. skin cancer)
    'lung',     #Chronic lung disease such as chronic bronchitis or emphysema
    'heart',    #Heart attack, coronary heart disease, angina, congestive heart 
                #failure, or other heart problems
    'strok',    #Stroke or cerebral vascular disease
    'arthr',    #Arthritis or rheumatism
    
    #self-reported health
    'shlt',
    
    #Depression
    'cesd',     #CES-D score
    
    #Frailty
    'weight',   #Self-reported weight
    'bmi',      #Self-reported BMI
    'height',   #Self-reported height
    'grp',      #hand grip strength maximum measurement
    'timwlk',   #timed walk test time
    'ltactx',   #Freq light phys activ
    'mdactx',   #Freq moderate phys activ
    'vgactx',   #Freq vigorous phys activ
    'effort',   #CESD: Everything an effort
    'going',    #CESD: Could not get going

    ###Complicating Features
    'adl6h',    # bathe, dress, and eat, getting in/out of bed and walking 
                #across a room and using the toilet
    'iadl5a',   #using the phone, managing money, and taking medications, 
                #shopping for groceries and preparing hot meals
    'homcar',   #in the last two years, has any medically-trained person 
                #come to your home to help you? MEDICALLY-TRAINED PERSONS 
                #INCLUDE PROFESSIONAL NURSES, VISITING NURSE'S AIDES, 
                #PHYSICAL OR OCCUPATIONAL THERAPISTS, CHEMOTHERAPISTS, 
                #RESPIRATORY OXYGEN THERAPISTS, AND HOSPICE CAREGIVERS
    'psyche', #
    'inhptn', #Total number of helpers ever helped
    'hlppdtn',  #HLPPDTN: Total number of helpers paid to help
    'hsptim', #RwHSPTIM is the reported number of hospital stays

    #### Utilization
    #'hsptim', #RwHSPTIM is the reported number of hospital stays
    'hspnit',    #Hospitl nghts, prv 12 mos    
    'nrsnit',    #nights in nurs home, prv 2 yrs
    'doctim',    #number of doctor visits, prv 12 mos
    'homcar',    #home hlth care, prv 2 yrs
    'outpt',     #outpatient surgry, prev 2 yrs
    'spcfac',    #Spec hlth facilty, prv 2 yrs
    'oopmd',     #Mexp Amt:Out of pkt med exp, prv 2 yrs
    
    #### Insurance
    'higov',      #R was covered by Gov plan Categ
    'govmr',      #R had Gov plan-Medicare Categ
    'govmd',      #R has Gov plan-Medicaid Categ
    'govva',      #R had Gov plan-CHAMPUS/CHAMPVA Categ
    
}

#### Define prefix per wave

wave_prefix = {
     8:'r8',
     9:'r9',
    10:'r10',
    11:'r11',
    12:'r12',
    13:'r13',
    14:'r14',
}

#### Define columns with a wave-specific pre-fix, for which a lag-variable is needed

prev_wave_variables = {
    'weight',
}

#### Define lag prefix per wave

prev_wave_prefix = {
     8:'r7', 
     9:'r8',
    10:'r9',
    11:'r10',
    12:'r11',
    13:'r12',
    14:'r13',
}

#### Define function to extract and transform wave-specific data

def extract_transform_randhrs(wave):
    
    #Start with the full RAND HRS dataset containing data from 1992 to 2018
    df = pd.DataFrame(df_randhrs)
    
    #Define relevant variables
    selected_variables = (['hhidpn'] + ['ragender'] + ['racohbyr'] + 
                          [wave_prefix[wave] + 
                           variable for variable in variables] + 
                          [prev_wave_prefix[wave] + 
                           variable for variable in prev_wave_variables]) 
    
    #Reduce columns to include wave-specific columns required for global 
    #impressions segmentation
    df = df[df.columns.intersection(selected_variables)]
    
    #Reformat hhidpn variable
    df.hhidpn = df.hhidpn.astype(int) #remove decimal point for hhidpn 
    df.hhidpn = df.hhidpn.astype(object) #convert from numeric to a string
    df.hhidpn = df.hhidpn.astype(str).str.zfill(9) #fill the left with zero's 
                                                   #to make hhidpn 9 digits
    
    #Rename lag variables
    lhs_prev = ([prev_wave_prefix[wave] + 
                 variable for variable in prev_wave_variables]) 
    rhs_prev = ['prev_' + variable for variable in prev_wave_variables] 
    fin_var_prev = len(prev_wave_variables)
    #rename columns
    for x in range(0, fin_var_prev):
        df.rename(columns = {lhs_prev[x]:rhs_prev[x]}, inplace = True) 
    
    #Rename gender
    df.rename(columns = {'ragender':'gender'}, inplace = True) #rename columns
    
    #Remove wave specific prefixes on individual level
    lhs = [wave_prefix[wave] + variable for variable in variables] 
    rhs = [variable for variable in variables] 
    fin_var = len(variables)
    for x in range(0, fin_var):
        df.rename(columns = {lhs[x]:rhs[x]}, inplace = True) #rename columns
       
    #Add wave indicator
    df['wave'] = wave
    
    return df 

#### Extract and transform dataset per wave

df_8_a = extract_transform_randhrs(8)
df_9_a = extract_transform_randhrs(9)
df_10_a = extract_transform_randhrs(10)
df_11_a = extract_transform_randhrs(11)
df_12_a = extract_transform_randhrs(12)
df_13_a = extract_transform_randhrs(13)
df_14_a = extract_transform_randhrs(14)

print("Each wave should have the same number of columns extracted")
print(len(df_8_a.columns), " columns extracted from wave 8")
print(len(df_9_a.columns), " columns extracted from wave 9")
print(len(df_10_a.columns), " columns extracted from wave 10")
print(len(df_11_a.columns), " columns extracted from wave 11")
print(len(df_12_a.columns), " columns extracted from wave 12")
print(len(df_13_a.columns), " columns extracted from wave 13")
print(len(df_14_a.columns), " columns extracted from wave 14")

#### Identify columns missing from first wave (wave 8)

print(df_8_a.columns.symmetric_difference(df_9_a.columns), 
      " <- columns missing from wave 8")

#### Identify columns missing from first wave (wave 14)

print(df_14_a.columns.symmetric_difference(df_9_a.columns), 
      " <- columns missing from wave 14")

### Define "df_randhrs_2" as stacked dataframes

df_randhrs_2 = [df_8_a, 
               df_9_a, 
               df_10_a, 
               df_11_a, 
               df_12_a, 
               df_13_a, 
               df_14_a]
df_randhrs_2 = pd.concat(df_randhrs_2)

print("Each wave should have same amount of entries")
print(df_randhrs_2["wave"].value_counts().sort_index())

#  3. Cross-Wave Tracker File

This section extracts and transforms the __Cross-Wave Tracker File__.

#### Load tracker data as "df_trk" and format column headings

df_trk = pd.read_stata(path_trk)
df_trk.columns= df_trk.columns.str.lower()

#### Define columns that have a wave-specific pre-fix

variables = {
    'iwmonth', #Month that interview took place
    'iwyear',  #Year that interview took place
    'iwtype',  #Wave X Whether Interviewed in the Wave
    'alive',   # interview status per wave
    'pmwgtr',  #final respondent weight for wave's physical measures subsample
    'couple',  #married or living with a partner as if married), the variable 
               #xCOUPLE is equal to 1. Otherwise, it is 5.
}

#### Define prefix per wave

wave_prefix = {
     8:'k',
     9:'l',
    10:'m',
    11:'n',
    12:'o',
    13:'p',
    14:'q',
    15:'r',
}

#### Define columns that to not have a wave-specific pre-fix and convert set to a list

selected_fixed_variables = {
    'hhidpn',             #unique id
    'birthmo',            #year of death
    'birthyr',            #month of death
    'knowndeceasedmo',    #best information about death date 
    'knowndeceasedyr',    #best information about death date 
    'lastalivemo',        #month on which we last knew the respondent was alive
    'lastaliveyr',        #year on which we last knew the respondent was alive
    'phymsr04',           #sample indicator for participation in physical 
                          #measures section
    'eftfassign',         #The variable EFTFASSIGN holds the respondents’ 
                          #permanent assignment for enhanced face-to-face 
                          #rotation from 2006 onward.
}

selected_fixed_variables = list(selected_fixed_variables)

#### Define function to extract and transform wave-specific data

def extract_transform_trk(wave):
    
    #Start with the full tracker dataset containing data from 1992 to 2018
    df = pd.DataFrame(df_trk)
    
    #Construct unique ID by combining household and personal ID
    df["hhidpn"] = df["hhid"] + df["pn"]
    
    #Define relevant variables with wave-specific prefixes 
    selected_variables = ['hhidpn'] + [wave_prefix[wave] + 
                                       variable for variable in variables]

    #Reduce columns to include wave-specific columns required 
    df_wave_specific = df[df.columns.intersection(selected_variables)]
    df_wave_specific = pd.DataFrame(df_wave_specific)
    
    #Remove wave specific prefixes
    df_wave_specific.columns = df_wave_specific.columns.str[1:]
    df_wave_specific.rename(columns={"hidpn": "hhidpn"}, inplace = True)
    
    #Reduce columns to include non-wave-specific columns required 
    df_non_wave_specific = df[df.columns.intersection(selected_fixed_variables)]
    df_non_wave_specific = pd.DataFrame(df_non_wave_specific)
    
    #Merge dynamic and static variables
    df = pd.merge(df_wave_specific, df_non_wave_specific, 
                  how="outer", on="hhidpn")
    
    #Add wave indicator
    df['wave'] = wave
    
    return df

#### Extract and transform dataset per wave

df_8_b = extract_transform_trk(8)
df_9_b = extract_transform_trk(9)
df_10_b = extract_transform_trk(10)
df_11_b = extract_transform_trk(11)
df_12_b = extract_transform_trk(12)
df_13_b = extract_transform_trk(13)
df_14_b = extract_transform_trk(14)
df_15_b = extract_transform_trk(15)

print("Each wave should have the same number of columns extracted (exept wave 15)")
print(len(df_8_b.columns), " columns extracted from wave 8")
print(len(df_9_b.columns), " columns extracted from wave 9")
print(len(df_10_b.columns), " columns extracted from wave 10")
print(len(df_11_b.columns), " columns extracted from wave 11")
print(len(df_12_b.columns), " columns extracted from wave 12")
print(len(df_13_b.columns), " columns extracted from wave 13")
print(len(df_14_b.columns), " columns extracted from wave 14")
print(len(df_15_b.columns), " columns extracted from wave 15") 
print("The weights are not yet available for the preliminary data",
      " released in wave 15")

#### Identify columns missing from first wave (wave 8)

print(df_8_b.columns.symmetric_difference(df_9_b.columns), 
      " <- columns missing from wave 8")

#### Identify columns missing from second last wave (wave 14)

print(df_14_b.columns.symmetric_difference(df_9_b.columns), 
      " <- columns missing from wave 14")

#### Identify columns missing from last wave (wave 15) 

print(df_15_b.columns.symmetric_difference(df_9_b.columns), 
      " <- columns missing from wave 15")
print("Anticipate pmwgtr to be missing")

### Define "df_trk_2" as stacked dataframes

df_trk_2 = [df_8_b, 
            df_9_b, 
            df_10_b, 
            df_11_b, 
            df_12_b, 
            df_13_b, 
            df_14_b,
            df_15_b]

df_trk_2 = pd.concat(df_trk_2)

print("Each wave should have same amount of entries")
df_trk_2["wave"].value_counts().sort_index() 

### Define "df_trk_3" as additional tracking file of persons who died in wave 15

df_trk_3 = pd.DataFrame(df_15_b)
df_trk_3 = df_trk_3[df_trk_3.alive==5]

# 4. RAND HRS Fat Files

Extract and transform __2018 RAND HRS Fat File__, __2016 RAND HRS Fat File__, __2014 RAND HRS Fat File__, __2012 RAND HRS Fat File__, __2010 RAND HRS Fat File__, __2008 RAND HRS Fat File__, __2006 RAND HRS Fat File__.

In the RAND HRS dataset, the cognition variables and self-reported taking of prescription medications are taken directly from the HRS imputations of cognitive functioning. These imputations are calculated
only for the final release data, meaning that there are no cognitive variables released for Wave 14 yet. We therefore extract all cognition variables from the HRS Core dataset to ensure consistency across waves. 

#### Define prefix per wave

# Value: prefix
wave_prefix = {
    14: "Q",
    13: "P",
    12: "O",
    11: "N",
    10: "M",
    9: "L",
    8: "K",
}

#### Define required variables

variables = {
    'D174',    #Immediate word recall
    'D184',    #Delayed word recall
    'D142',    #Serial 7 test - 1
    'D143',    #Serial 7 test - 2
    'D144',    #Serial 7 test - 3
    'D145',    #Serial 7 test - 4
    'D146',    #Serial 7 test - 5
    'D155',    #Object naming test: scissors
    'D156',    #Object naming test: cactus
    'G097',    #Suppose in the future, you needed help with basic personal care 
               #activities like eating or dressing. Do you have relatives or 
               #friends (besides your [husband/wife/partner]) who would be 
               #willing and able to help you over a long period of time?
    
    #Reported taking medication for
    'C006',    # high blood pressure or hypertension
    'C011',    # treat or control your [diabetes/blood sugar]
    'C032',    # treatment for your lung condition
    'C037',    # taking or carrying medication for your heart problem
    'C042',    # taking or carrying medication because of your heart attack
    'C046',    # taking or carrying medications because of angina or chest pain
    'C050',    # taking or carrying any medication for congestive heart failure
    'C282',    # In order to regulate your heart rhythm are you now 
               #taking any medication
    'C060',    # taking any medications because of your stroke or its complications 
    'N360',    # To help lower your cholesterol?
    'N361',    # For pain in your joints or muscles?
    'N362',    # For asthma or allergies or other breathing problems?
    'N363',    # For stomach problems?
    'N364',    # To help you sleep?
    'N365',    # To help relieve anxiety or depression?
    'N283',    # Do you regularly take prescription medications other than aspirin 
               #to thin your blood or to prevent blood clots?
    
    #Out of pocket payments
    'N106',    # hospital care
    'N119',    # nursing home care
    'N139',    # out of hospital surgery
    'N156',    # doctor visits
    'N168',    # dental care
    'N180',    # prescriptions
    'N194',    # in-home health care
    'N328',    # hospice care (exit interview only)
    'N239',    # other services
}

#### Define function to extract and transform wave-specific data

def extract_and_transform_HSR_Core(wave):
    
    #Load wave-specific dataset
    df = pd.read_stata(global_path + path_hrsCore[wave])
    
    #Define as dataframe
    df = pd.DataFrame(df)
    
    #Harmonize heading cases across waves
    df.columns = df.columns.str.upper() # convert all headings to uppercase 
                                        # (necessary for wave: 6,8 and 9)
    
    #Define relevant variables with wave-specific prefixes
    selected_variables = (["HHIDPN"]+
                          [wave_prefix[wave] +
                           variable for variable in variables]) 
    
    #Reduce columns to include wave-specific columns required 
    df = df[df.columns.intersection(selected_variables)] 
    
    #Remove wave specific prefixes
    df.columns = df.columns.str[1:]
    df.rename(columns={"HIDPN": "hhidpn"}, inplace = True)
    
    #Reformat hhidpn variable
    df.hhidpn = df.hhidpn.astype(int) #remove decimal point for hhidpn 
    df.hhidpn = df.hhidpn.astype(object) #convert from numeric to a string
    df.hhidpn = df.hhidpn.astype(str).str.zfill(9) #fill the left with zero's 
                                                   #to make hhidpn 9 digits
        
    #Add wave indicator
    df['wave'] = wave

    return df

df_8_c = extract_and_transform_HSR_Core(8)
df_9_c = extract_and_transform_HSR_Core(9)
df_10_c = extract_and_transform_HSR_Core(10)
df_11_c = extract_and_transform_HSR_Core(11)
df_12_c = extract_and_transform_HSR_Core(12)
df_13_c = extract_and_transform_HSR_Core(13)
df_14_c = extract_and_transform_HSR_Core(14)

print("Each wave should have the same number of columns extracted")
print(len(df_8_c.columns), " columns extracted from wave 8")
print(len(df_9_c.columns), " columns extracted from wave 9")
print(len(df_10_c.columns), " columns extracted from wave 10")
print(len(df_11_c.columns), " columns extracted from wave 11")
print(len(df_12_c.columns), " columns extracted from wave 12")
print(len(df_13_c.columns), " columns extracted from wave 13")
print(len(df_14_c.columns), " columns extracted from wave 14")

### Define "df_hrsCore" as stacked dataframes

df_hrsCore = [df_8_c, 
              df_9_c, 
              df_10_c, 
              df_11_c, 
              df_12_c, 
              df_13_c, 
              df_14_c]

df_hrsCore = pd.concat(df_hrsCore)

print("Each wave should have a different amount of entries since the rows ",
      "reflects the actual number of participants in each wave")
df_hrsCore["wave"].value_counts().sort_index() 

# 5. Merge dataframes

This section merges df_randhrs_2, df_trk_2, and df_hrsCore into a single dataframe

#### Define testing variables to evaluate merge

df_randhrs_2["a"] = 1
df_trk_2["b"] = 1
df_hrsCore["c"] = 1

#### Merge df_randhrs_2 with df_hsrCore

Perform a left merge on df_randhrs_2 and df_hrsCore. Keep all entries from df_randhrs_2.

data0 = pd.merge(df_randhrs_2, df_hrsCore, how="left", on=["hhidpn", "wave"])
        
#Test whether all entries consists of data from both df_randhrs_2 and df_trk_2
a = int(data0['a'].sum())
c = int(data0['c'].sum())
print("Merged dataset contains ", a, " rows from df_randhrs_2")
print("Merged dataset contains ", c, " rows from df_hsrCore")

#Test whether (1) all rows from df_randhrs_2 are in the merged dataset
#AND (2) there is at least one row from df_hsrCore in the merged dataset
if (a == len(df_randhrs_2)) & (c > 0):
    print("Rows merged correctly")
else:
    print("Rows not merged correctly")

#Check if columns have merged correctly
cols_tot = ([col for col in df_hrsCore.columns] + 
            [col for col in df_randhrs_2.columns])
cols_tot = np.array(cols_tot)
cols_tot = len(np.unique(cols_tot))
cols_merged = len(df_hrsCore.columns) + len(df_randhrs_2.columns) - 2

#Test if all columns from both datasets are present in the merged dataset
if cols_tot == cols_merged:
    print("Columns merged correctly")
else:
    print("Columns not merged correctly")

#### Merge data0 with df_trk_2

Perform a left merge on data0 and df_trk_2. Keep all entries from data0.

#Merge data0 with df_hsrCore. Keep all entries  
data = pd.merge(data0, df_trk_2, how="left", on=["hhidpn", "wave"])

#Test whether all entries consists of data from both df_randhrs_2 and df_trk_2
a = int(data["a"].value_counts().sort_index())
b = int(data["b"].value_counts().sort_index())

#Confirm that the same number of entries are present across each wave
print("Number of rows in df_randhrs_2:", data["a"].sum())
print("Number of rows in df_trk_2:", data["b"].sum())
print("Number of rows in merged dataframe", len(data))

#Confirm that the merge did not lose information
if (max(a, b) == len(data)) & (min(a, b) > 0):
    print("Rows merged correctly")
else:
    print("Rows not merged correctly")    

#Check if columns have merged correctly
cols_tot = ([col for col in df_hrsCore.columns] + 
            [col for col in df_randhrs_2.columns] + 
            [col for col in df_trk_2.columns])
cols_tot = np.array(cols_tot)
cols_tot = len(np.unique(cols_tot))
cols_merged = (len(df_hrsCore.columns) + 
               len(df_randhrs_2.columns) - 2 + 
               len(df_trk_2.columns) - 2)

#Test if all columns from both datasets are present in the merged dataset
if cols_tot == cols_merged:
    print("Columns merged correctly")
else:
    print("Columns not merged correctly")

# 6. Clean data

#### Evaluate starting number of entries per wave

n_data_0 = data["wave"].value_counts().sort_index()
print("Number of entries per wave:")
print(n_data_0)

#### Remove entries for whom living status is not known for given wave

data.dropna(subset = ["alive"], inplace=True) 
    
n_data_1 = data["wave"].value_counts().sort_index()
n_dropped_1 = n_data_0 - n_data_1
print("Number of entries per wave:")
print(n_data_1)
print(" ")
print("Number of entries dropped per wave:")
print(n_dropped_1)

#### Remove entries if their death was recorded in a previous wave

data = data[data.alive != 6]

n_data_2 = data["wave"].value_counts().sort_index()
n_dropped_2 = n_data_1 - n_data_2
print("Number of entries per wave:")
print(n_data_2)
print(" ")
print("Number of entries dropped per wave:")
print(n_dropped_2)

#### Remove entries if no contact was made in the given wave

data = data[data.alive != 2]

n_data_3 = data["wave"].value_counts().sort_index()
n_dropped_3 = n_data_2 - n_data_3
print("Number of entries per wave:")
print(n_data_3)
print(" ")
print("Number of entries dropped per wave:")
print(n_dropped_3)

#### Remove respondents who are recorded as dead in the first interview wave that they participated in

#Drop all respondents that are reported as dead in wave 8
n_data_5 = len(data)
               
#sort individuals per id, per wave
data_alive = pd.DataFrame(data)
data_alive = data_alive.sort_values(['hhidpn', 'wave'])

#isolate first interview per ID
data_alive = data_alive.groupby('hhidpn').first().reset_index()
data_alive = pd.DataFrame(data_alive)

#drop all entries that are not alive
data_alive = data_alive[data_alive.alive==1]

#define series as all IDs in data
s = pd.Series(data.hhidpn)

#identify which of the IDs have a live first observation
s = s.isin(pd.Series(data_alive.hhidpn))

#restrict the data to only individuals that have at least one live observation
data = data[s]
data = pd.DataFrame(data)

n_data_6 = len(data)

#Evaluate entries removed per wave
n_dropped_6 = n_data_5 - n_data_6
print("Number of entries per wave:")
print(data["wave"].value_counts().sort_index())
print("Number of entries dropped:")
print(n_dropped_6)

data.alive.value_counts().sort_index()

#### Remove entries from respondents younger than 50

#assign random birth month if birth month is not known
data['randmo'] = [np.random.randint(1, 12) for i in range(0, len(data))] 
data['birthmo_2'] = data.birthmo
data.loc[(data.birthmo == 0), 'birthmo_2'] = data.randmo

#calculate time at birth
data["birthtime"] = ((data.birthmo_2-1)/12+data.birthyr)

#calculate time of interview
data["iwtime"] = ((data.iwmonth-1)/12+data.iwyear)

#isolate wave and time of interview 
data_iw_m = data[['wave', 'iwtime']]
data_iw_m = pd.DataFrame(data_iw_m)

#Calculate the mean interview date per wave
data_iw_m = data_iw_m.groupby('wave')[['iwtime']].mean()
data_iw_m = data_iw_m.reset_index()

#rename variable
data_iw_m.rename(columns = {'iwtime':'iwtime_mean'}, inplace = True)

#Insert mean interview date per wave to data
data = pd.merge(data, data_iw_m, how="left", on=["wave"])
data.loc[pd.isnull(data.iwtime), 'iwtime'] = data.iwtime_mean

#Calculate age at time of interview
data['age'] = data.iwtime - data.birthtime


n_data_7 = data["wave"].value_counts().sort_index()
#drop all entries that are not alive
data = data[data.age>=50]

n_data_8 = data["wave"].value_counts().sort_index()
n_dropped_8 = n_data_7 - n_data_8
print("Number of entries per wave:")
print(n_data_8)
print(" ")
print("Number of entries dropped per wave:")
print(n_dropped_8)

#### Remove entries where respondent did not complete at least one physical measure

n_data_9 = data["wave"].value_counts().sort_index()

#Remove individuals who did not participate in the 
#physical measures section of the survey
data = data[(data.pmwgtr>0)|(data.alive==5)]


#Evaluate entries removed per wave
n_data_10 = data["wave"].value_counts().sort_index()
n_dropped_10 = n_data_9 - n_data_10
print("Number of entries per wave:")
print(n_data_10)
print("Number of entries dropped per wave:")
print(n_dropped_10)

# 7. Extraction Statistics

print("Observations per wave:")
print(data.wave.value_counts().sort_index())
print(" ")

n_total = len(data)
print(n_total, " <- total number of observations in dataset")

n_interview = len(data[data.alive!=5])
print(n_interview, " <- total enhanced face-to-face interviews")

n_deceased = len(data[data.alive==5])
print(n_deceased, " <- total participants deceased")

n_uniqueID = len(data.hhidpn.unique())
print(n_uniqueID, " <- number of individuals in study")

# 8. Export to csv

### Write "dataHRS.csv"

Write dataset to be used in 02_Global_Impressions_Segmentation.ipynb and 03_Complicating_Features.ipynb

filepath_dataHRS = os.path.join(global_path2, "dataHRS.csv")

data.to_csv(filepath_dataHRS, index=False)

### Write "df_trk_2.csv"

Write to dataset to be used in 04_SurvivalAnalysis.ipynb

filepath = os.path.join(global_path2, "df_trk_2.csv")

df_trk_2.to_csv(filepath, index=False)
